Feature/db

+++ app/src/main/java/com/example/running_app/data/database/dao/TB_GPS.java
@@ -12,15 +12,16 @@ import androidx.room.PrimaryKey;
                 parentColumns = "run_id",
                 childColumns = "run_id",
                 onDelete = ForeignKey.CASCADE))
+//@Entity(tableName = "TB_GPS")
 public class TB_GPS {
     @PrimaryKey(autoGenerate = true)
     private int gps_id;
     @ColumnInfo(name = "run_id")
     private int run_id;
     @ColumnInfo(name = "lat")
-    private long lat;
+    private double lat;
     @ColumnInfo(name = "lon")
-    private long lon;
+    private double lon;
     @ColumnInfo(name = "create_at")
     private String create_at;
 
@@ -40,19 +41,19 @@ public class TB_GPS {
         this.run_id = run_id;
     }
 
-    public long getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(long lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 
-    public long getLon() {
+    public double getLon() {
         return lon;
     }
 
-    public void setLon(long lon) {
+    public void setLon(double lon) {
         this.lon = lon;
     }
 
@@ -63,6 +64,4 @@ public class TB_GPS {
     public void setCreate_at(String create_at) {
         this.create_at = create_at;
     }
-
-
 }
